🗂️ Fold Queues admin into System → Jobs as tabs
Merge the queue drain config + tail UI into the existing system jobs page
behind a Jobs/Queues tab (deep-linkable via ?tab=queues) and remove the
standalone system.queues route and nav entry.
