+++ platform/scms/app/routes/app/system/menu.ts
@@ -53,12 +53,6 @@ export function buildMenu(baseUrl: string, userScopes: string[]) {
           url: `${baseUrl}/jobs`,
         },
         {
-          name: 'admin.queues',
-          label: 'Queues',
-          icon: 'radio',
-          url: `${baseUrl}/queues`,
-        },
-        {
           name: 'admin.design',
           label: 'Design',
           url: `${baseUrl}/design`,
